Add files via upload
A restaurant has many customers and a manager. A customer can initiate 1 or many orders. An order can be special order or a normal order.
A Customer has various characteristics, Name, Ages, Address and his characteristics cannot be access by others but by setters and getters.
An Order has properties date of type Date, Reference No and we apply the same characteristics of the of the customer.
Special order has a particular property called Special Name and inherits of order. A normal order has a particular property called normal name and inherits of order.